% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalibration.R
\name{doCalibration}
\alias{doCalibration}
\title{POINT AND CLICK CALIBRATION CODE}
\usage{
doCalibration(path = ".", nsim = 999, trim = FALSE)
}
\arguments{
\item{path}{the location on disk where the calibration folders are
initially searched for (the default where the directory selection
dialogue opens).}

\item{nsim}{number of simulations to run (see \code{\link{calibration}})}
}
\description{
Function to provide an easy interface to use the calibation
methods in the loggercal package
}
\details{
When running the function the user will be asked to set various
directorise where internal and external calibration data are stored:
internalCalDir = the filepath of where the internal calibration file is located
externalCalDir = the filepath of where the external calibration file is located which
                 is appropriate for the internal calibration that you are working on
}
\examples{

\dontrun{
# set up a demo calibration folder structure
# please set this to a suitable location on your system
path <- "D:\\\\temp\\\\Calibration_data"
caldir <- file.path(path, "Logger_calibrations", "250416")
ukasdir <- file.path(path, "External_Calibrations", "010915")
dir.create(caldir, recursive = TRUE)
dir.create(ukasdir, recursive = TRUE)
file.copy(system.file("calibration_files", "FullCalibration_250416.csv", package = "loggercal"), caldir)
file.copy(system.file("calibration_files", "UKASCalibration_319151.csv", package = "loggercal"), ukasdir)
 doCalibration(path, nsim = 10)
}

}
