% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findLag}
\alias{findLag}
\title{Add together two numbers.}
\usage{
findLag(internalCal, lagtry = -10:10)
}
\arguments{
\item{lagtry}{A number.}

\item{cal}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

