% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getCoefs}
\alias{getCoefs}
\title{Add together two numbers.}
\usage{
getCoefs(reverseCalMod)
}
\arguments{
\item{reverseCalMod}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

