% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggercal-package.R
\docType{package}
\name{loggercal-package}
\alias{loggercal}
\alias{loggercal-package}
\title{Tools for calibrating field temperature loggers}
\description{
Contains all functions required to (1) Read in temperature logger calibration data
  (External and Internal) (2) calculate approximate calibration functions to convert
  raw temps to true and raw temps to SE of approx (3) output a file containing calibration
  results by logger for use as a bulk upload to FLEObs database
}
\author{
Colin Millar.
}

