% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myFilters}
\alias{myFilters}
\title{FILTERS FOR CHOOSEING CALIBRATION FILES.}
\format{An object of class \code{matrix} with 4 rows and 2 columns.}
\usage{
myFilters
}
\description{
note only available on windows.
}
\keyword{datasets}

