% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.InternalCal.R
\name{print.InternalCal}
\alias{print.InternalCal}
\title{Add together two numbers.}
\usage{
\method{print}{InternalCal}(x, ...)
}
\arguments{
\item{x}{A number.}

\item{...}{Numeric, complex, or logical vectors.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
