% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalCal_funcs.R
\name{readExternalCal}
\alias{readExternalCal}
\title{Add together two numbers.}
\usage{
readExternalCal(dirname)
}
\arguments{
\item{dirname}{The directory containing the UKAS calibration files.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

