% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalCal_funcs.R
\name{readInternalCal}
\alias{readInternalCal}
\title{Add together two numbers.}
\usage{
readInternalCal(fname)
}
\arguments{
\item{fname}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

