% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCalPlot.R
\name{saveCalPlot}
\alias{saveCalPlot}
\title{Save calibration plot as a png file}
\usage{
saveCalPlot(filename, plotFunc, width = 17, height = 17, ppi = 1200)
}
\arguments{
\item{filename}{A number.}

\item{plotFunc}{A function which returns a plot}

\item{width}{A number.}

\item{height}{A number.}

\item{ppi}{A number.}
}
\description{
Description
}
\examples{
1 + 1
}
