% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs_funcs.R
\name{tableCoefs}
\alias{tableCoefs}
\title{Add together two numbers.}
\usage{
tableCoefs(coefs, cal)
}
\arguments{
\item{coefs}{A number.}

\item{cal}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
