% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Add together two numbers.}
\usage{
calibration(internalCal, externalCalMod, n = 99)
}
\arguments{
\item{internalCal}{A number.}

\item{externalCalMod}{A number.}

\item{n}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

