% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Fit a calibration regression and estimate logger error.}
\usage{
calibration(internalCal, externalCalMod, nsim = 999)
}
\arguments{
\item{internalCal}{A number.}

\item{externalCalMod}{A number.}

\item{nsim}{number of simulations to run in the estimation.}
}
\value{
models for calibration and SE models.
}
\description{
This function:
   * identifies calibration loggers
   * chooses one calibration logger to be used
   * finds the best lag
   * fits a calibration model to each non-calibration logger
   * returns models for calibration and SE models
}
