% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doCalibration.R
\name{doCalibration}
\alias{doCalibration}
\title{Run a full calibration}
\usage{
doCalibration(nsim = 9999)
}
\arguments{
\item{nsim}{The number of simulations to use to evaluate confidence intervals.}
}
\description{
This function does everything from read in files, select calibtation period and
thens runs a calibration and saves the file.
}
\examples{
1 + 1
}

