% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{findStartSec}
\alias{findStartSec}
\title{Add together two numbers.}
\usage{
findStartSec(cal)
}
\arguments{
\item{cal}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
