% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{lagmod}
\alias{lagmod}
\title{Add together two numbers.}
\usage{
lagmod(x, y, lag = 0, formula = y ~ poly(x, 2))
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{lag}{A number.}

\item{formula}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

