% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myFilters}
\alias{myFilters}
\title{FILTERS FOR CHOOSEING CALIBRATION FILES.}
\format{\preformatted{ chr [1:4, 1:2] "UKAS calibration" "Internal calibration" "all calibration" "all files" "UKAS*.csv" "full*.csv" "*.csv" ...
}}
\usage{
myFilters
}
\description{
note only available on windows.
}
\keyword{datasets}

