% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\docType{data}
\name{myFilters}
\alias{myFilters}
\title{Add together two numbers.}
\format{\preformatted{ chr [1:4, 1:2] "UKAS calibration" "Internal calibration" "all calibration" "all files" "UKAS*.csv" "full*.csv" "*.csv" ...
}}
\usage{
myFilters
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
\examples{
1 + 1
}
\keyword{datasets}

