% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalCal_funcs.R
\name{plot.InternalCal}
\alias{plot.InternalCal}
\title{Add together two numbers.}
\usage{
\method{plot}{InternalCal}(x, type = "scaled", xlim = NULL, ...)
}
\arguments{
\item{x}{A number.}

\item{type}{A number.}

\item{xlim}{A number.}

\item{...}{other graphics parameters to pass on to plotting commands.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

