% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalCal_funcs.R
\name{readExternalCal}
\alias{readExternalCal}
\title{Add together two numbers.}
\usage{
readExternalCal(fnames)
}
\arguments{
\item{fnames}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}

