% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalCal_funcs.R
\name{readInternalCal}
\alias{readInternalCal}
\title{Add together two numbers.}
\usage{
readInternalCal(dirname)
}
\arguments{
\item{dirname}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
