% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{savePlot}
\alias{savePlot}
\title{Save calibration plot as a png file}
\usage{
savePlot(filename, plotFunc, width = 17, height = 17, ppi = 1200)
}
\arguments{
\item{filename}{A number.}

\item{plotFunc}{A function which returns a plot}

\item{width}{A number.}

\item{height}{A number.}

\item{ppi}{A number.}
}
\description{
Description
}
\examples{
1 + 1
}

