% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Add together two numbers.}
\usage{
trim(x, ...)
}
\arguments{
\item{x}{A number.}

\item{...}{Numeric, complex, or logical vectors.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
